:truck: Move call's logic to another file.
